+++ FishMMO-Unity/Assets/Scripts/Client/UI/Controls/Login/UILoadingScreen.cs
@@ -124,7 +124,7 @@ namespace FishMMO.Client
 		/// </summary>
 		/// <param name="attempts">The current attempt number.</param>
 		/// <param name="maxAttempts">The maximum number of allowed attempts.</param>
-		public void Client_OnReconnectAttempt(byte attempts, byte maxAttempts)
+		public void Client_OnReconnectAttempt(int attempts, int maxAttempts)
 		{
 			LoadingImage.sprite = DefaultLoadingScreenSprite;
 			Show();
